Move listener handler to a dedicated packaged and also included its unit test
